requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Configuration
pydantic-settings==2.1.0

# AI Integration
requests==2.31.0

# File Processing (optional - for advanced features)
beautifulsoup4==4.12.3
python-docx==1.1.0
PyPDF2==3.0.1

# Environment
python-dotenv==1.0.0
